Refactor for modularity
